requirements.txt
Django==3.2.15
psycopg2==2.9.3
yahoo-fin==0.8.9.1
djangorestframework==3.13.1
celery==5.2.7
django-celery-results==2.4.0
django-celery-beat==2.3.0
redis==4.3.4
channels==3.0.5
channels-redis==3.4.1
simplejson==3.17.6